Avoid backreference inside a character class (fixes #4)
The author of this regexp is actually @sergeypererva. Kudos, man!

The problem with the original regexp is that it uses something that a
regexp shouldn't normally contain: [^\1] is not a valid thing.

More information: http://www.regular-expressions.info/backref.html
(See section "Parentheses and Backreferences Cannot Be Used Inside
Character Classes")