features/string-conversion.feature: add a failing test
Although the new regexp makes all the tests pass, it doesn't cover the
case, which was previously covered by the original regexp. This test
proves the statement. If you run the test suit with the old regexp, it
should still contain only one, the same old, failure: "Convert string
with the number 1 in it". However this test will pass. On the contrary,
the new regexp can't please this test.